Add repository and fix model
