+++ src/main/java/com/classhub/api/model/Students.java
@@ -1,15 +1,14 @@
-package com.classhub.model;
+package com.classhub.api.model;
 
 import jakarta.persistence.*;
-
 import lombok.Getter;
 import lombok.Setter;
 
 @Getter
 @Setter
 @Entity
-@Table(name = "Teachers")
-public class Teachers {
+@Table(name = "Students")
+public class Students {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     @Column(name = "id", nullable = false)
@@ -30,6 +29,9 @@ public class Teachers {
     @Column(name = "pwd")
     private String pwd;
 
+    @Column(name = "cluster")
+    private String cluster;
 
-
+    @Column(name = "faculty")
+    private String faculty;
 }
