+++ src/main/java/com/classhub/api/model/Teachers.java
@@ -1,14 +1,15 @@
-package com.classhub.model;
+package com.classhub.api.model;
 
 import jakarta.persistence.*;
+
 import lombok.Getter;
 import lombok.Setter;
 
 @Getter
 @Setter
 @Entity
-@Table(name = "Students")
-public class Students {
+@Table(name = "Teachers")
+public class Teachers {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     @Column(name = "id", nullable = false)
@@ -29,9 +30,6 @@ public class Students {
     @Column(name = "pwd")
     private String pwd;
 
-    @Column(name = "group")
-    private String group;
 
-    @Column(name = "faculty")
-    private String faculty;
+
 }
