+++ src/main/java/com/classhub/api/repository/AdministratorsRepository.java
@@ -1,0 +1,7 @@
+package com.classhub.api.repository;
+
+import com.classhub.api.model.Administrators;
+import org.springframework.data.jpa.repository.JpaRepository;
+
+public interface AdministratorsRepository extends JpaRepository<Administrators,Long> {
+}
