+++ src/main/java/com/classhub/api/repository/StudentsRepository.java
@@ -1,0 +1,7 @@
+package com.classhub.api.repository;
+
+import com.classhub.api.model.Students;
+import org.springframework.data.jpa.repository.JpaRepository;
+
+public interface StudentsRepository extends JpaRepository<Students,Long> {
+}
